Create an AudioProcessor to process audio and add a method to get the duration of a track when the file is given

+++ backend/song-service/src/main/java/edu/personal/grailsfm/songservice/service/impl/TrackServiceImpl.java
@@ -19,7 +19,7 @@ public class TrackServiceImpl implements TrackService {
 
     @Override
     public Object createTrack(CreateTrackDto trackDto) throws UnsupportedAudioFileException, IOException {
-        Integer duration = audioProcessor.calculateDurationOfTrack(trackDto.file());
+        Float duration = audioProcessor.calculateDurationOfTrack(trackDto.file());
         return null;
     }
 }
